+++ src/base/hip/hip_matrix_bcsr.cpp
@@ -114,10 +114,10 @@ namespace rocalution
     }
 
     template <typename ValueType>
-    void HIPAcceleratorMatrixBCSR<ValueType>::AllocateBCSR(int nnzb,
-                                                           int nrowb,
-                                                           int ncolb,
-                                                           int blockdim)
+    void HIPAcceleratorMatrixBCSR<ValueType>::AllocateBCSR(int64_t nnzb,
+                                                           int     nrowb,
+                                                           int     ncolb,
+                                                           int     blockdim)
     {
         assert(nnzb >= 0);
         assert(ncolb >= 0);
@@ -172,8 +172,13 @@ namespace rocalution
     }
 
     template <typename ValueType>
-    void HIPAcceleratorMatrixBCSR<ValueType>::SetDataPtrBCSR(
-        int** row_offset, int** col, ValueType** val, int nnzb, int nrowb, int ncolb, int blockdim)
+    void HIPAcceleratorMatrixBCSR<ValueType>::SetDataPtrBCSR(int**       row_offset,
+                                                             int**       col,
+                                                             ValueType** val,
+                                                             int64_t     nnzb,
+                                                             int         nrowb,
+                                                             int         ncolb,
+                                                             int         blockdim)
     {
         assert(*row_offset != NULL);
         assert(*col != NULL);
